Move model types to `types`, move resp types to `protocol`.
