+++ lowlevel/api/app.go
@@ -1,22 +1,12 @@
 package api
 
-type AppVersionInfo struct {
-	CreatedTime int64  `json:"created_time"`
-	VersionCode string `json:"version_code"`
-	VersionUrl  string `json:"version_url"`
-}
-
-type AppVersionResult struct {
-	Android    AppVersionInfo `json:"android"`
-	LinuxApp   AppVersionInfo `json:"linux_115"`
-	MacBrowser AppVersionInfo `json:"mac"`
-	MacApp     AppVersionInfo `json:"mac_115"`
-	WinBrowser AppVersionInfo `json:"win"`
-	WinApp     AppVersionInfo `json:"window_115"`
-}
+import (
+	"github.com/deadblue/elevengo/internal/protocol"
+	"github.com/deadblue/elevengo/lowlevel/types"
+)
 
 type AppVersionSpec struct {
-	_JsonpApiSpec[AppVersionResult, StandardResp]
+	_JsonpApiSpec[types.AppVersionResult, protocol.StandardResp]
 }
 
 func (s *AppVersionSpec) Init() *AppVersionSpec {
