+++ lowlevel/api/image.go
@@ -1,18 +1,9 @@
 package api
 
-type ImageGetResult struct {
-	FileName string `json:"file_name"`
-	FileSha1 string `json:"file_sha1"`
-	Pickcode string `json:"pick_code"`
-
-	SourceUrl string   `json:"source_url"`
-	OriginUrl string   `json:"origin_url"`
-	ViewUrl   string   `json:"url"`
-	ThumbUrls []string `json:"all_url"`
-}
+import "github.com/deadblue/elevengo/lowlevel/types"
 
 type ImageGetSpec struct {
-	_StandardApiSpec[ImageGetResult]
+	_StandardApiSpec[types.ImageGetResult]
 }
 
 func (s *ImageGetSpec) Init(pickcode string) *ImageGetSpec {
