+++ lowlevel/api/upload.go
@@ -7,9 +7,10 @@ import (
 	"time"
 
 	"github.com/deadblue/elevengo/internal/crypto/hash"
+	"github.com/deadblue/elevengo/internal/protocol"
 	"github.com/deadblue/elevengo/internal/util"
 	"github.com/deadblue/elevengo/lowlevel/client"
-	"github.com/deadblue/elevengo/lowlevel/errors"
+	"github.com/deadblue/elevengo/lowlevel/types"
 )
 
 const (
@@ -19,31 +20,8 @@ const (
 	UploadMaxSizeSample = 200 * 1024 * 1024
 )
 
-type UploadInfoResult struct {
-	UserId  int
-	UserKey string
-}
-
-//lint:ignore U1000 This type is used in generic.
-type _UploadInfoResp struct {
-	_BasicResp
-
-	UserId  int    `json:"user_id"`
-	UserKey string `json:"userkey"`
-}
-
-func (r *_UploadInfoResp) Extract(v any) error {
-	if ptr, ok := v.(*UploadInfoResult); !ok {
-		return errors.ErrUnsupportedResult
-	} else {
-		ptr.UserId = r.UserId
-		ptr.UserKey = r.UserKey
-	}
-	return nil
-}
-
 type UploadInfoSpec struct {
-	_JsonApiSpec[UploadInfoResult, _UploadInfoResp]
+	_JsonApiSpec[types.UploadInfoResult, protocol.UploadInfoResp]
 }
 
 func (s *UploadInfoSpec) Init() *UploadInfoSpec {
@@ -112,77 +90,8 @@ type UploadInitParams struct {
 	SignValue string
 }
 
-type UploadInitResult struct {
-	Exists bool
-	// Upload parameters
-	Oss struct {
-		Bucket      string
-		Object      string
-		Callback    string
-		CallbackVar string
-	}
-	SignKey   string
-	SignCheck string
-}
-
-//lint:ignore U1000 This type is used in generic.
-type _UploadInitResp struct {
-	Request   string `json:"request"`
-	Version   string `json:"version"`
-	ErrorCode int    `json:"statuscode"`
-	ErrorMsg  string `json:"statusmsg"`
-
-	Status   int    `json:"status"`
-	PickCode string `json:"pickcode"`
-
-	// New fields in upload v4.0
-	SignKey   string `json:"sign_key"`
-	SignCheck string `json:"sign_check"`
-
-	// OSS upload fields
-	Bucket   string `json:"bucket"`
-	Object   string `json:"object"`
-	Callback struct {
-		Callback    string `json:"callback"`
-		CallbackVar string `json:"callback_var"`
-	} `json:"callback"`
-
-	// Useless fields
-	FileId   int    `json:"fileid"`
-	FileInfo string `json:"fileinfo"`
-	Target   string `json:"target"`
-}
-
-func (r *_UploadInitResp) Err() error {
-	// Ignore 701 error
-	if r.ErrorCode == 0 || r.ErrorCode == 701 {
-		return nil
-	}
-	return errors.ErrUnexpected
-}
-
-func (r *_UploadInitResp) Extract(v any) (err error) {
-	if ptr, ok := v.(*UploadInitResult); !ok {
-		err = errors.ErrUnsupportedResult
-	} else {
-		switch r.Status {
-		case 1:
-			ptr.Oss.Bucket = r.Bucket
-			ptr.Oss.Object = r.Object
-			ptr.Oss.Callback = r.Callback.Callback
-			ptr.Oss.CallbackVar = r.Callback.CallbackVar
-		case 2:
-			ptr.Exists = true
-		case 7:
-			ptr.SignKey = r.SignKey
-			ptr.SignCheck = r.SignCheck
-		}
-	}
-	return
-}
-
 type UploadInitSpec struct {
-	_JsonApiSpec[UploadInitResult, _UploadInitResp]
+	_JsonApiSpec[types.UploadInitResult, protocol.UploadInitResp]
 }
 
 func (s *UploadInitSpec) Init(params *UploadInitParams, helper *UploadHelper) *UploadInitSpec {
@@ -209,43 +118,8 @@ func (s *UploadInitSpec) Init(params *UploadInitParams, helper *UploadHelper) *U
 	return s
 }
 
-type UploadTokenResult struct {
-	AccessKeyId     string
-	AccessKeySecret string
-	SecurityToken   string
-	Expiration      time.Time
-}
-
-//lint:ignore U1000 This type is used in generic.
-type _UploadTokenResp struct {
-	StatusCode      string `json:"StatusCode"`
-	AccessKeyId     string `json:"AccessKeyId"`
-	AccessKeySecret string `json:"AccessKeySecret"`
-	SecurityToken   string `json:"SecurityToken"`
-	Expiration      string `json:"Expiration"`
-}
-
-func (r *_UploadTokenResp) Err() error {
-	if r.StatusCode == "200" {
-		return nil
-	}
-	return errors.ErrUnexpected
-}
-
-func (r *_UploadTokenResp) Extract(v any) error {
-	if ptr, ok := v.(*UploadTokenResult); !ok {
-		return errors.ErrUnsupportedResult
-	} else {
-		ptr.AccessKeyId = r.AccessKeyId
-		ptr.AccessKeySecret = r.AccessKeySecret
-		ptr.SecurityToken = r.SecurityToken
-		ptr.Expiration, _ = time.Parse(time.RFC3339, r.Expiration)
-	}
-	return nil
-}
-
 type UploadTokenSpec struct {
-	_JsonApiSpec[UploadTokenResult, _UploadTokenResp]
+	_JsonApiSpec[types.UploadTokenResult, protocol.UploadTokenResp]
 }
 
 func (s *UploadTokenSpec) Init() *UploadTokenSpec {
@@ -253,46 +127,8 @@ func (s *UploadTokenSpec) Init() *UploadTokenSpec {
 	return s
 }
 
-type UploadSampleInitResult struct {
-	Host        string
-	Object      string
-	Callback    string
-	AccessKeyId string
-	Policy      string
-	Signature   string
-}
-
-//lint:ignore U1000 This type is used in generic.
-type _UploadSampleInitResp struct {
-	Host        string `json:"host"`
-	Object      string `json:"object"`
-	Callback    string `json:"callback"`
-	AccessKeyId string `json:"accessid"`
-	Policy      string `json:"policy"`
-	Signature   string `json:"signature"`
-	Expire      int64  `json:"expire"`
-}
-
-func (r *_UploadSampleInitResp) Err() error {
-	return nil
-}
-
-func (r *_UploadSampleInitResp) Extract(v any) error {
-	if ptr, ok := v.(*UploadSampleInitResult); !ok {
-		return errors.ErrUnsupportedResult
-	} else {
-		ptr.Host = r.Host
-		ptr.Object = r.Object
-		ptr.Callback = r.Callback
-		ptr.AccessKeyId = r.AccessKeyId
-		ptr.Policy = r.Policy
-		ptr.Signature = r.Signature
-	}
-	return nil
-}
-
 type UploadSampleInitSpec struct {
-	_JsonApiSpec[UploadSampleInitResult, _UploadSampleInitResp]
+	_JsonApiSpec[types.UploadSampleInitResult, protocol.UploadSampleInitResp]
 }
 
 func (s *UploadSampleInitSpec) Init(userId string, fileName string, fileSize int64, target string) *UploadSampleInitSpec {
@@ -304,19 +140,8 @@ func (s *UploadSampleInitSpec) Init(userId string, fileName string, fileSize int
 	return s
 }
 
-type UploadSampleResult struct {
-	AreaId     util.IntNumber `json:"aid"`
-	CategoryId string         `json:"cid"`
-	FileId     string         `json:"file_id"`
-	FileName   string         `json:"file_name"`
-	FileSize   util.IntNumber `json:"file_size"`
-	FileSha1   string         `json:"sha1"`
-	PickCode   string         `json:"pick_code"`
-	CreateTime util.IntNumber `json:"file_ptime"`
-}
-
 type UploadSampleSpec struct {
-	_StandardApiSpec[UploadSampleResult]
+	_StandardApiSpec[types.UploadSampleResult]
 	payload client.Payload
 }
