+++ lowlevel/types/shortcut.go
@@ -1,0 +1,11 @@
+package types
+
+type ShortcutInfo struct {
+	FileId   string `json:"file_id"`
+	FileName string `json:"file_name"`
+	Sort     string `json:"sort"`
+}
+
+type ShortcutListResult struct {
+	List []*ShortcutInfo `json:"list"`
+}
