+++ lowlevel/types/void.go
@@ -1,0 +1,4 @@
+package types
+
+// VoidResult describes a void result.
+type VoidResult struct{}
